requirements.txt
asgiref==3.3.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.7
django-admin-search==0.3.6
django-braces==1.14.0
django-cleanup==5.1.0
django-import-export==2.5.0
django-maintenance-mode==0.16.0
django-pwa==1.0.10
django-session-security==2.6.6
django-storages==1.11.1
et-xmlfile==1.0.1
google-api-core==1.26.3
google-auth==1.28.1
google-cloud-core==1.6.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
idna==2.10
MarkupPy==1.14
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.9
Pillow==8.1.2
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.4
xlrd==2.0.1
xlwt==1.3.0
